refactored to experiment name, created TrainerPL
