+++ 03-multiclass/src/trainers/__init__.py
@@ -1,1 +1,2 @@
 from .class_trainer import TrainerClass
+from .pl_trainer import TrainerPL
